Removed a type mismatch/conversion warning
git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@12859 01821946-851b-0410-b7f4-f1477c70e57a

+++ io/cfile.cpp
@@ -75,7 +75,7 @@ strcpy (szFolder, szMainFolder);
 if (*szFolder)
 	AppendSlash (szFolder);
 strcat (szFolder, szSubFolder);
-if ((l = strlen (szFolder))) {
+if ((l = (int) strlen (szFolder))) {
 	if (*szFilter) {
 		AppendSlash (szFolder);
 		strcat (szFolder, szFilter);
